% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getIMEXVersions}
\alias{getIMEXVersions}
\title{Get IMEX versions included in the msigdb package}
\usage{
getIMEXVersions()
}
\value{
a character, stating the IMEX versions available in this package.
}
\description{
Get IMEX versions included in the msigdb package
}
\examples{
getIMEXVersions()

}
