% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getMsigdbVersions}
\alias{getMsigdbVersions}
\title{Get MSigDB versions included in the msigdb package}
\usage{
getMsigdbVersions()
}
\value{
a character, stating the versions available in this package.
}
\description{
Get MSigDB versions included in the msigdb package
}
\examples{
getMsigdbVersions()

}
