% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMSigInfo.R
\name{getMSigInfo}
\alias{getMSigInfo}
\title{Get information about the gene set}
\usage{
getMSigInfo(Msig)
}
\arguments{
\item{Msig}{name of the molecular signature}
}
\description{
Navigates to the relevant web page about the molecular signature
}

