% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{hcop}
\alias{hcop}
\title{Mouse to Human homolog map from HCOP}
\format{
a data.frame, with mouse and human Entrez IDs.
}
\usage{
hcop
}
\description{
Entrez ID mapping between mouse and human homologs from the HCOP database.
}
\references{
Wright, M. W., Eyre, T. A., Lush, M. J., Povey, S., & Bruford, E.
A. (2005). HCOP: the HGNC comparison of orthology predictions search tool.
Mammalian genome, 16(11), 827-828.
}
\keyword{datasets}
