% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessMSigDB.R
\name{listSubCollections}
\alias{listSubCollections}
\title{List all sub-collection types within a MSigDB gene set collection}
\usage{
listSubCollections(gsc)
}
\arguments{
\item{gsc}{a GeneSetCollection object, storing GeneSets from the MSigDB}
}
\value{
a character vector, containing character codes for all
sub-collections present in the GeneSetCollection object.
}
\description{
This function lists all the sub-collection types present in a MSigDB gene set
collection. Descriptions of sub-collections can be found at the MSigDB
website.
}
\examples{
gsc = getMsigdb('hs', 'SYM')
listSubCollections(gsc)

}
