% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{msigdbR-package}
\alias{msigdbR-package}
\title{The Molecular Signatures Database (MSigDB)}
\description{
This ExperimentHub package contains gene expression signatures from the
molecular signatures database (MSigDB). The molecular signatures database
(MSigDB) is a collection of over 25000 gene expression signatures that are
grouped into collections and sub-collections. Metadata associated with
signatures is collected and stored in the data in this package.
}
\details{
Data in this package are stored in a GeneSetCollection object from the
GSEABase package. Each gene expression signature is stored in a GeneSet
object from the GSEABase package.

The following datasets are included in this package:
\enumerate{
\item msigdb.v7.1.SYM - The MSigDB v7.1 with gene expression signatures defined
using gene symbols.
\item msigdb.v7.1.EZID - The MSigDB v7.1 with gene expression signatures defined
using Entrez IDs.
}
}
