% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessMSigDB.R
\name{subsetCollection}
\alias{subsetCollection}
\title{Title}
\usage{
subsetCollection(gsc, collection, subcollection = NULL)
}
\arguments{
\item{gsc}{a GeneSetCollection object, storing GeneSets from the MSigDB}

\item{collection}{}

\item{subcollection}{}
}
\value{

}
\description{
Title
}
