% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addKEGG.R
\name{appendKEGG}
\alias{appendKEGG}
\title{Add KEGG pathway derived gene sets to a collection}
\usage{
appendKEGG(gsc)
}
\arguments{
\item{gsc}{a GeneSetCollection object, storing GeneSets from the MSigDB}
}
\value{
a GeneSetCollection object, storing gene sets from the MSigDB
including the downloaded KEGG gene sets.
}
\description{
This function adds gene sets derived from KEGG pathways to the MSigDB data
stored in this package. Direct user-end download from the MSigDB is required
to ensure KEGG licenses are adhered to.
}
\examples{
library(GSEABase)

gs1 <- GeneSet(setName = 'gs1', geneIdType = SymbolIdentifier())
gsc <- GeneSetCollection(gs1)
gsc <- appendKEGG(gsc)
gsc

}
