% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addKEGG.R
\name{appendKEGG}
\alias{appendKEGG}
\title{Add KEGG pathway derived gene sets to a collection}
\usage{
appendKEGG(gsc, id = c("sym", "ezid"), version = "7.1")
}
\arguments{
\item{gsc}{a GeneSetCollection object, storing GeneSets from the MSigDB}

\item{id}{a character, either 'sym' or 'ezid' representing the gene
identifier to download (symbol or entrez id respectively).}

\item{version}{a character, stating the version of msigdb to download
(>=7.1).}
}
\value{
a GeneSetCollection object, storing gene sets from the MSigDB
including the downloaded KEGG gene sets.
}
\description{
This function adds gene sets derived from KEGG pathways to the MSigDB data
stored in this package. Direct user-end download from the MSigDB is required
to ensure KEGG licenses are adhered to.
}
\examples{
library(msigdb)
library(GSEABase)

gsc <- msigdb.hs.v7.1.SYM()
gsc <- appendKEGG(gsc, id = 'sym', version = '7.1')
gsc

}
