% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addKEGG.R
\name{appendKEGG}
\alias{appendKEGG}
\title{Add KEGG pathway derived gene sets to a collection}
\usage{
appendKEGG(gsc, version = getMsigdbVersions())
}
\arguments{
\item{gsc}{a GeneSetCollection object, containing MSigDB genesets in the form
of GeneSet objects.}

\item{version}{a character, stating the version of MSigDB to be retrieved
(should be >= 7.2). See \code{getMsigdbVersions()}.}
}
\value{
a GeneSetCollection object, storing gene sets from the MSigDB
including the downloaded KEGG gene sets.
}
\description{
This function adds gene sets derived from KEGG pathways to the MSigDB data
stored in this package. Direct user-end download from the MSigDB is required
to ensure KEGG licenses are adhered to.
}
