% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMSigInfo.R
\docType{data}
\name{MSigDB}
\alias{MSigDB}
\title{MSigDB signature sets}
\format{An object of class \code{list} of length 8.}
\usage{
MSigDB
}
\description{
This object houses gene lists from Broad Institute Molecular Signatures
}
\keyword{datasets}

